feat: create ConvertToPointDataValue function
